+++ collatex-pythonport/collatex/__init__.py
@@ -4,7 +4,7 @@
 __name__ = 'CollateX'
 __author__ = 'Ronald Haentjens Dekker'
 __email__ = 'ronald.dekker@huygens.knaw.nl'
-__version__ = '2.1.3rc2'
+__version__ = '2.2'
 
 
 from collatex.core_classes import Collation